Adds functionality to the POST /symfile endpoint
This commit adds functionality for the POST /symfile endpoint, which
saves the Symfiles on the filesystem and makes a database entry for them
this currently only checks if it starts with "MODULE", it does not yet
check if the transmitted file already exists in the database. It also
uses information from the first line of the Symfile to fill database
fields such as Os and Arch.