Adding better API routes
This continues work on #20 although it is still a work in progress.
The product endpoint should now be fully functional including GETting
a product or a list of all, POSTing a new product, PUTting an updated
version online (thinking about changing the verb to PATCH) and DELETEing
a product. If there is an error, the "error" field in the JSON will be
non-null, else the saved object will be returned in the "object" field,
of course, for a successful delete both fields are null as there is no
object to return anymore.

Working on the methods for Versions and Users too.