First complete API version
The first version of the JSON API is now finished,
admin actions (users, versions, products) include GETing,
POSTing, PUTing and DELETEing. All the actions take JSON
body which looks like the actual object you are working with
although without *At, the ID is server generated on POSTing or
supplied via the route on GETting, PUTting and DELETEing.

There is also now an 'API' for the frontend now, which just
returns the content used to render the HTML template as a
JSON object. This could change in the future if a JavaScript-
based frontend is implemented. The routes for these are

```
/api/stats
/api/crashes
/api/crashes/:id
/api/reports
/api/reports/:id
```

In general it is also possible to use the front-end routes as
the returned content is based on the Accept header at the moement,
but I would recommend to use the `/api` routes in case the output
gets more API-specific some day (including adding other verbs like
DELETE).

This fixes #20