Add button to mark crashes as fixed
Fixed crashes won't show up in the Crashes overview anymore as long
as there is no report with a Version which is not yet part of the Crash.
If the latter occurs the crash will be marked as un-fixed and be shown again.
The detail page is always visible and also allows you to un-fix a Crash
by clicking a button.

Fixes #19