Adding GET endpoints for crashes, reports and symfiles
Filtering is possible with filter_field=fieldname and filter_value=value
It is also possible to AND-combine multiple filters by setting the query-parameters
multiple times.

Ordering is possible with order_field=fieldname and order_direction=desc (the default
ordering direction is asc so there is no parameter needed). If direction is something other
than `desc` it will also be ordered ascending.

Pagination is possible with limit=numer and offset=number. If nothing is specified limit
will be 50 and offset 0.