Show precise value for short uptimes
A process may crash at start-up, i.e. within the first 5 seconds.
In these cases, it is useful to be able to tell the difference between
e.g. 50 ms and 500 ms, or 1.2s and 1.8s. So we show millisecond values
for uptimes smaller than 5 seconds.

Print the correct value for h (when h > 24) while we're at it.